feat: shard _type_protocols.py into 6 domain-grouped files and migrate consumers
Split monolithic core/_type_protocols.py (29 protocols) into 6 shards:
- _type_protocols_logging (7): AuditLog, TokenLog, TimingLog, McpResponseLog,
  GitHubApiLog, SupportsDebug, SupportsLogger
- _type_protocols_execution (4): TestRunner, HeadlessExecutor, OutputPatternResolver,
  WriteExpectedResolver
- _type_protocols_github (3): GitHubFetcher, CIWatcher, MergeQueueWatcher
- _type_protocols_workspace (5): WorkspaceManager, CloneManager, SessionSkillManager,
  SkillLister, SkillResolver
- _type_protocols_recipe (4): RecipeRepository, MigrationService, DatabaseReader,
  ReadOnlyResolver
- _type_protocols_infra (6): GateState, BackgroundSupervisor, FleetLock,
  QuotaRefreshTask, TokenFactory, CampaignProtector

Migrate all consumers: types.py (6 star-imports), _type_helpers.py (SkillResolver),
execution/{ci,github,merge_queue,session_log}.py (TYPE_CHECKING GitHubApiLog).

Add SKILL_FILE_ADVISORY_MAP and SKILL_ACTIVATE_DEPS_REQUIRED to core/__init__.pyi.

Update arch tests: EXEMPTIONS core/ 27→32, protocol collection loop, whitelist key.
Update cascade tests: replace _type_protocols with 6 shard stems in universal modules.
Add tests/core/test_type_protocol_shards.py with __all__ and runtime_checkable checks.
