+++ src/autoskillit/core/__init__.pyi
@@ -160,7 +160,9 @@ from .types import SESSION_TYPE_FLEET as SESSION_TYPE_FLEET
 from .types import SESSION_TYPE_LEAF as SESSION_TYPE_LEAF
 from .types import SESSION_TYPE_ORCHESTRATOR as SESSION_TYPE_ORCHESTRATOR
 from .types import SESSION_TYPE_ORDER as SESSION_TYPE_ORDER
+from .types import SKILL_ACTIVATE_DEPS_REQUIRED as SKILL_ACTIVATE_DEPS_REQUIRED
 from .types import SKILL_COMMAND_PREFIX as SKILL_COMMAND_PREFIX
+from .types import SKILL_FILE_ADVISORY_MAP as SKILL_FILE_ADVISORY_MAP
 from .types import SKILL_TOOLS as SKILL_TOOLS
 from .types import SOUS_CHEF_L2_SECTIONS as SOUS_CHEF_L2_SECTIONS
 from .types import SOUS_CHEF_MANDATORY_SECTIONS as SOUS_CHEF_MANDATORY_SECTIONS
