+++ tests/_test_filter.py
@@ -120,7 +120,12 @@ _CORE_UNIVERSAL_MODULES: frozenset[str] = frozenset(
         "logging",
         "types",
         "_type_constants",
-        "_type_protocols",
+        "_type_protocols_logging",
+        "_type_protocols_execution",
+        "_type_protocols_github",
+        "_type_protocols_workspace",
+        "_type_protocols_recipe",
+        "_type_protocols_infra",
         "_type_enums",
         "_type_subprocess",
         "_type_results",
