+++ tests/arch/test_python_no_hardcoded_temp.py
@@ -29,7 +29,7 @@ _TEMP_PATH_WHITELIST: dict[str, str] = {
     "recipe/repository.py": "validate_from_path default arg mirrors canonical default",
     # Justification: RecipeRepository protocol method includes temp_dir_relpath with
     # the canonical default value for backward compatibility.
-    "core/_type_protocols.py": "RecipeRepository.validate_from_path default arg",
+    "core/_type_protocols_recipe.py": "RecipeRepository.validate_from_path default arg",
     # Justification: SkillsDirectoryProvider default ctor arg matches canonical default.
     "workspace/session_skills.py": "SkillsDirectoryProvider default arg",
     # Justification: docstring example referencing the canonical default path.
