+++ tests/contracts/test_protocol_definitions.py
@@ -1,6 +1,6 @@
-"""Tests for Protocol definitions in core/_type_protocols.py.
+"""Tests for Protocol definitions in core/_type_protocols_*.py shards.
 
-REQ-PROTO-007: SkillLister Protocol must live in core/_type_protocols.py.
+REQ-PROTO-007: SkillLister Protocol must live in core/_type_protocols_workspace.py.
 """
 
 from __future__ import annotations
@@ -9,7 +9,7 @@ import inspect
 
 
 def test_github_fetcher_has_update_issue_body() -> None:
-    from autoskillit.core._type_protocols import GitHubFetcher
+    from autoskillit.core._type_protocols_github import GitHubFetcher
 
     assert hasattr(GitHubFetcher, "update_issue_body")
     sig = inspect.signature(GitHubFetcher.update_issue_body)
@@ -20,7 +20,7 @@ def test_github_fetcher_has_update_issue_body() -> None:
 
 
 def test_github_fetcher_no_add_comment() -> None:
-    from autoskillit.core._type_protocols import GitHubFetcher
+    from autoskillit.core._type_protocols_github import GitHubFetcher
 
     assert not hasattr(GitHubFetcher, "add_comment"), (
         "add_comment must be removed after all call sites are migrated"
@@ -29,10 +29,10 @@ def test_github_fetcher_no_add_comment() -> None:
 
 def test_skill_lister_protocol_defined() -> None:
     """REQ-PROTO-007: SkillLister Protocol must live in
-    core/_type_protocols.py and define a `list_all() -> list[Any]`
+    core/_type_protocols_workspace.py and define a `list_all() -> list[Any]`
     method, so L2 recipe code can type-annotate against L0 instead of
     binding to the L1 workspace concrete class."""
-    from autoskillit.core._type_protocols import SkillLister
+    from autoskillit.core._type_protocols_workspace import SkillLister
 
     assert hasattr(SkillLister, "list_all")
     sig = inspect.signature(SkillLister.list_all)
@@ -40,7 +40,7 @@ def test_skill_lister_protocol_defined() -> None:
 
 
 def test_skill_resolver_satisfies_skill_lister() -> None:
-    from autoskillit.core._type_protocols import SkillLister
+    from autoskillit.core._type_protocols_workspace import SkillLister
     from autoskillit.workspace.skills import DefaultSkillResolver
 
     instance: SkillLister = DefaultSkillResolver()
