+++ tests/recipe/test_repository.py
@@ -202,7 +202,7 @@ def test_recipe_repository_protocol_find_return_type_is_recipe_info() -> None:
     """
     import inspect
 
-    from autoskillit.core._type_protocols import RecipeRepository
+    from autoskillit.core._type_protocols_recipe import RecipeRepository
 
     sig = inspect.signature(RecipeRepository.find)
     ann = sig.return_annotation
