+++ tests/test_test_filter_core_cascade.py
@@ -24,7 +24,12 @@ class TestCoreUniversalModules:
             "logging",
             "types",
             "_type_constants",
-            "_type_protocols",
+            "_type_protocols_logging",
+            "_type_protocols_execution",
+            "_type_protocols_github",
+            "_type_protocols_workspace",
+            "_type_protocols_recipe",
+            "_type_protocols_infra",
             "_type_enums",
             "_type_subprocess",
             "_type_results",
